Implement user account management mechanism
Created a mechanism that manages user accounts. The mechanism can create accounts, log in to them, update user data, and retrieve it. Data are stored in an SQLite database. Access to accounts is managed through a token, which functions as a login-password combo, as well as a unique user identifier in the database.