Create Python and Godot modules to make it easier to use the new protocol
The protocol was created to meet all the needs of this project. It is simple and universal - it will be suitable for almost any type of project. It uses JSON and is based on three rules: type specifies the type of message, command indicates the action being performed, and payload contains additional data needed to execute the command.